Add 2 new buttons in dataCenter

+++ storage/commands/ping.js
@@ -1,11 +1,16 @@
 const Discord = require("discord.js");
 const config = require("../config.json");
 
-module.exports.run = async (bot, message, dataSpecialChannel)=>{
-    message.delete();
-    message.channel.send("pong")
+module.exports.run = async (bot, message, channel) => {
+    if (message) {
+        channel = message.channel;
+        if (message.deletable) message.delete();
+    }
+    channel.send("pong")
         .then(msg => {
-            msg.delete({ timeout: 10000 })
+            msg.delete({
+                timeout: 10000
+            })
         });
 };
